PoorsManTwitter/requirements.txt
appdirs==1.4.4
asgiref==3.2.3
astroid==2.4.2
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.5.3
beautifulsoup4==4.9.3
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
Cython==0.29.16
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
Django==3.0.4
django-allauth==0.44.0
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-fernet-fields==0.6
django-filter==2.2.0
django-freeze==0.6.4
django-js-asset==1.2.2
django-simple-history==3.0.0
django-softdelete==0.9.2
django-sslserver==0.22
djangorestframework==3.11.0
djangorestframework-simplejwt==4.7.0
filelock==3.0.12
findspark==1.4.2
getmac==0.8.2
idna==2.9
imageio==2.8.0
imutils==0.5.3
install==1.3.4
isort==4.3.21
joblib==0.14.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.6.3
matplotlib==3.2.1
mccabe==0.6.1
more-itertools==8.2.0
networkx==2.4
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.2.0.34
packaging==20.3
PassportEye==1.4.1
paypalrestsdk==1.13.1
pdfminer3k==1.3.2
Pillow==7.1.1
pluggy==0.13.1
ply==3.11
psycopg2==2.8.4
py==1.8.1
py4j==0.10.9
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.5.3
pyOpenSSL==19.1.0
pyotp==2.6.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
pyspark==3.1.2
pytesseract==0.3.3
pytest==5.4.1
python-dateutil==2.8.1
python-gettext==4.0
python3-openid==3.2.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.23.0
requests-oauthlib==1.3.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.3.1
stripe==2.48.0
tesseract==0.1.3
toml==0.10.1
twilio==6.42.0
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.2.0
virtualenv==20.4.4
wcwidth==0.1.9
wrapt==1.12.1
xmltodict==0.12.0
youtube-dl==2021.6.6
